LWWSet

Documentation

data LWWSet a

The LWW-Element-Set datastructure

toSet :: Ord a => LWWSet a -> Set a

Reduce LWWSet to a normal Set which contains only non-deleted elements

merge :: Ord a => LWWSet a -> LWWSet a -> LWWSet a

Here we use the fact that LWWSet is a Monoid

empty :: Ord a => LWWSet a

Construct an empty LWWSet, also from Monoid instance

unit :: Ord a => a -> TimeStamp -> LWWSet a

Construct an LWWSet with a single element

query :: Ord a => LWWSet a -> a -> Bool

Check if the element is in the LWWSet, which means it was added the latest add timestamp is greater that the last removal if any. Here we are biased to removal.

insert :: Ord a => LWWSet a -> a -> TimeStamp -> LWWSet a

Add new element to the LWWSet

remove :: Ord a => LWWSet a -> a -> TimeStamp -> LWWSet a

Remove an element from the LWWSet